feat(verification): Implement Combined Verification Strategy (Task 16.3)
This commit introduces the 'Combined' verification mode, allowing for flexible
contract verification. This strategy attempts abstract (axiom-based) verification
first, falling back to instantiation-based verification if the abstract pass
fails or is inconclusive.

Key changes include:
- Extended  with an optional  field
  to specify per-contract verification strategies.
- Updated  to parse , ,
  and  pragmas and populate the AST accordingly.
- Modified  to propagate the
  from AST assertions to s.
- Refactored  to iterate through individual verification
  conditions, applying their specified mode (AbstractOnly, InstantiationOnly,
  or Combined), or falling back to a function-level default.
- Ensured 's  carries the mode,
  and  populates it during VC generation.
- Added a comprehensive test case () to
   to validate all three verification modes and their
  interactions with axioms and function parameters.
- Addressed various compilation errors and warnings to ensure a clean build.

This completes Task 16.3 of the Aether V2 syntax migration, significantly
enhancing the static analysis capabilities for generic contracts.

+++ src/debug/dwarf.rs
@@ -443,7 +443,7 @@ impl DwarfGenerator {
             self.line_program.entries.push(LineEntry {
                 address: *block_id as u64 * 0x1000, // Placeholder address
                 file: file_index,
-                line: (*block_id + 1) as u32, // Placeholder line number
+                line: (*block_id + 1), // Placeholder line number
                 column: 1,
                 is_stmt: true,
                 basic_block: true,
@@ -455,7 +455,7 @@ impl DwarfGenerator {
                 self.line_program.entries.push(LineEntry {
                     address: (*block_id as u64 * 0x1000) + (stmt_idx as u64 * 4),
                     file: file_index,
-                    line: (*block_id + 1) as u32,
+                    line: (*block_id + 1),
                     column: (stmt_idx + 1) as u32,
                     is_stmt: true,
                     basic_block: false,
